Tidy up and refresh.
- Update to .NET SDK v6.0.428
- Remove reference to Mono
- Clarify `DebugTypeLogger`
- package.ps1: move build step to the top, throw on error
- README.md: mention testing performed on the latest version (v2026-01-5)

+++ package.ps1
@@ -1,14 +1,14 @@
 Set-StrictMode -Version Latest
 $ErrorActionPreference = "Stop"
 
+dotnet build --configuration "Release"
+if (-not $?) { throw "Build failed" }
+
 Remove-Item -Force -Recurse -ErrorAction "Continue" -Path "temp/"
 mkdir -Path "temp\package\IL2CPP\BepInEx\config","temp\package\IL2CPP\BepInEx\plugins\FuriousTare"
 Expand-Archive "BepInEx\BepInEx-Unity.IL2CPP-*.zip" "temp\package\IL2CPP"
 
 Copy-Item -Path "BepInEx\config\IL2CPP\BepInEx.cfg" -Destination "temp\package\IL2CPP\BepInEx\config\"
-
-dotnet build --configuration "Release"
-
 Copy-Item -Path "FuriousTareIL2CPP\bin\Release\net6.0\FuriousTareIL2CPP.dll" -Destination "temp\package\IL2CPP\BepInEx\plugins\FuriousTare"
 
 $Today = (Get-Date).ToUniversalTime().ToString('yyyy-MM-dd')
